// athena-plugin-factions/shared/interfaces.ts
type _id = string;

export const RankPermissionNames = {
    manageRanks: 'manageRanks',
    manageRankPermissions: 'manageRankPermissions',
    manageMembers: 'manageMembers',
    kickMembers: 'kickMembers',
    addMembers: 'addMembers',
};

/**
 * Permissions that belong to an individual rank.
 * @export
 * @interface RankPermissions
 */
export interface RankPermissions {
    /**
     * Can this rank manage ranks lower than itself?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    manageRanks: boolean;

    /**
     * Can this rank manage rank permissions lower than itself?
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    manageRankPermissions: boolean;

    /**
     * Can this rank manage members by promoting and demoting lower than itself?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    manageMembers: boolean;

    /**
     * Can this rank kick members lower than itself?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    kickMembers: boolean;

    /**
     * Can this rank add members lower than itself?
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    addMembers: boolean;

    /**
     * Can manage vehicles by purchasing vehicles available.
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    manageVehicles?: boolean;

    /**
     * Can open storages
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    canOpenStorages: boolean;
}

/**
 * Character reference for a faction.
 *
 * @export
 * @interface FactionCharacter
 */
export interface FactionCharacter {
    [key: string]: any;

    /**
     * The character ID of the Character.
     *
     * @type {string}
     * @memberof FactionCharacter
     */
    id: _id;

    /**
     * Name of the Character
     *
     * @type {string}
     * @memberof FactionCharacter
     */
    name: string;

    /**
     * A unique identifier for the rank.
     *
     * @type {string}
     * @memberof FactionCharacter
     */
    rank: string;

    /**
     * Does this member have Ownership of the Faction?
     *
     * @type {boolean}
     * @memberof FactionCharacter
     */
    hasOwnership: boolean;
}

/**
 * A rank inside of the Faction
 *
 * @export
 * @interface FactionRank
 */
export interface FactionRank {
    [key: string]: any;

    /**
     * A unique identifier for the Faction Rank
     * Should be auto-generated.
     *
     * @type {string}
     * @memberof FactionRank
     */
    uid?: string;

    /**
     * The name of this Rank
     *
     * @type {string}
     * @memberof FactonRank
     */
    name: string;

    /**
     * Custom action permissions this rank has access to.
     *
     * @type {Array<string>}
     * @memberof FactionRank
     */
    actionPermissions: Array<string>;

    /**
     * Permissions for the rank.
     *
     * @type {RankPermissions}
     * @memberof FactionRank
     */
    rankPermissions: RankPermissions;

    /**
     * Vehicles this rank has keys to by default.
     *
     * @type {Array<_id>}
     * @memberof FactionRank
     */
    vehicles: Array<_id>;

    /**
     * The higher the weight the more important they are.
     * Owner rank is always 99.
     * Nothing can exceed 99.
     *
     * @type {number}
     * @memberof FactionRank
     */
    weight: number;
}